Problem: too much going on in code cell
Solution: create an execute cell action

Problem: This is not pretty code, needs guidance from #123
