+++ src/components/Cell/CodeCell.jsx
@@ -10,6 +10,7 @@ import Display from 'react-jupyter-display-area';
 import {
   updateCellOutputs,
   updateCellExecutionCount,
+  executeCell,
 } from '../../actions';
 
 import {
@@ -36,49 +37,11 @@ export default class CodeCell extends React.Component {
     if (!e.shiftKey || e.key !== 'Enter') {
       return;
     }
-    const { iopub, shell } = this.context.channels;
 
-    if(!iopub || !shell) {
-      // TODO: handle attempt to execute when kernel not connected
-      return;
-    }
-
-    const executeRequest = createExecuteRequest(this.props.cell.get('source'));
-
-    // Limitation of the Subject implementation in enchannel
-    // we must shell.subscribe in order to shell.next
-    shell.subscribe(() => {});
-
-    // Set the current outputs to an empty list
-    this.context.dispatch(updateCellOutputs(this.props.id, new Immutable.List()));
-
-    const childMessages = iopub.childOf(executeRequest)
-                               .publish()
-                               .refCount();
-
-    childMessages.ofMessageType(['execute_input'])
-                 .pluck('content', 'execution_count')
-                 .first()
-                 .subscribe((ct) => {
-                   this.context.dispatch(updateCellExecutionCount(this.props.id, ct));
-                 });
-
-    // Handle all the nbformattable messages
-    childMessages
-         .ofMessageType(['execute_result', 'display_data', 'stream', 'error'])
-         .map(msgSpecToNotebookFormat)
-         // Iteratively reduce on the outputs
-         .scan((outputs, output) => {
-           return outputs.push(Immutable.fromJS(output));
-         }, new Immutable.List())
-         // Update the outputs with each change
-         .subscribe(outputs => {
-           this.context.dispatch(updateCellOutputs(this.props.id, outputs));
-         });
-
-    shell.next(executeRequest);
-
-    // TODO: Manage subscriptions
+    this.context.dispatch(executeCell(this.props.id,
+                                      this.props.cell.get('source'),
+                                      this.context.dispatch,
+                                      this.context.channels));
   }
 
   render() {
